[frontend] move delete button into update container (#10985)
