+++ opencti-platform/opencti-front/src/private/components/workspaces/WorkspaceEditionOverview.jsx
@@ -1,15 +1,23 @@
-import React, { Component } from 'react';
+import React from 'react';
 import * as PropTypes from 'prop-types';
 import { graphql, createFragmentContainer } from 'react-relay';
 import { Formik, Form, Field } from 'formik';
 import withStyles from '@mui/styles/withStyles';
 import { compose, pick } from 'ramda';
 import * as Yup from 'yup';
+import Button from '@mui/material/Button';
+import { Stack } from '@mui/material';
+import { useNavigate } from 'react-router-dom';
+import { WorkspacePopoverDeletionMutation } from './WorkspacePopover';
 import inject18n from '../../../components/i18n';
 import TextField from '../../../components/TextField';
 import { SubscriptionFocus } from '../../../components/Subscription';
 import { commitMutation } from '../../../relay/environment';
 import MarkdownField from '../../../components/fields/MarkdownField';
+import useDeletion from '../../../utils/hooks/useDeletion';
+import { deleteNode } from '../../../utils/store';
+import useApiMutation from '../../../utils/hooks/useApiMutation';
+import DeleteDialog from '../../../components/DeleteDialog';
 
 const styles = (theme) => ({
   drawerPaper: {
@@ -64,8 +72,17 @@ const workspaceValidation = (t) => Yup.object().shape({
   description: Yup.string().nullable(),
 });
 
-class WorkspaceEditionOverviewComponent extends Component {
-  handleChangeFocus(name) {
+const WorkspaceEditionOverviewComponent = ({ paginationOptions, t, workspace, context }) => {
+  const { id } = workspace;
+  const navigate = useNavigate();
+  const initialValues = pick(['name', 'description'], workspace);
+
+  const deletion = useDeletion({ handleClose: () => {} });
+  const { deleting, setDeleting, handleOpenDelete, handleCloseDelete } = deletion;
+
+  const [commit] = useApiMutation(WorkspacePopoverDeletionMutation);
+
+  const handleChangeFocus = (name) => {
     commitMutation({
       mutation: workspaceEditionOverviewFocus,
       variables: {
@@ -75,9 +92,9 @@ class WorkspaceEditionOverviewComponent extends Component {
         },
       },
     });
-  }
+  };
 
-  handleSubmitField(name, value) {
+  const handleSubmitField = (name, value) => {
     workspaceValidation(this.props.t)
       .validateAt(name, { [name]: value })
       .then(() => {
@@ -90,51 +107,87 @@ class WorkspaceEditionOverviewComponent extends Component {
         });
       })
       .catch(() => false);
-  }
+  };
 
-  render() {
-    const { t, workspace, context } = this.props;
-    const initialValues = pick(['name', 'description'], workspace);
-    return (
-      <Formik
-        enableReinitialize={true}
-        initialValues={initialValues}
-        validationSchema={workspaceValidation(t)}
-        onSubmit={() => true}
-      >
-        {() => (
-          <Form>
-            <Field
-              component={TextField}
-              name="name"
-              label={t('Name')}
-              fullWidth={true}
-              onFocus={this.handleChangeFocus.bind(this)}
-              onSubmit={this.handleSubmitField.bind(this)}
-              helperText={
-                <SubscriptionFocus context={context} fieldName="name" />
+  const submitDelete = () => {
+    const { type } = workspace;
+    setDeleting(true);
+    commit({
+      variables: { id },
+      updater: (store) => {
+        if (paginationOptions) {
+          deleteNode(store, 'Pagination_workspaces', paginationOptions, id);
+        }
+      },
+      onCompleted: () => {
+        setDeleting(false);
+        if (paginationOptions) {
+          handleCloseDelete();
+        } else if (type) {
+          navigate(`/dashboard/workspaces/${type}s`);
+        } else {
+          navigate('/dashboard');
+        }
+      },
+    });
+  };
+
+  return (
+    <Formik
+      enableReinitialize={true}
+      initialValues={initialValues}
+      validationSchema={workspaceValidation(t)}
+      onSubmit={() => true}
+    >
+      {() => (
+        <Form>
+          <Field
+            component={TextField}
+            name="name"
+            label={t('Name')}
+            fullWidth={true}
+            onFocus={handleChangeFocus}
+            onSubmit={handleSubmitField}
+            helperText={
+              <SubscriptionFocus context={context} fieldName="name" />
               }
-            />
-            <Field
-              component={MarkdownField}
-              name="description"
-              label={t('Description')}
-              fullWidth={true}
-              multiline={true}
-              rows="4"
-              style={{ marginTop: 20 }}
-              onFocus={this.handleChangeFocus.bind(this)}
-              onSubmit={this.handleSubmitField.bind(this)}
-              helperText={
-                <SubscriptionFocus context={context} fieldName="description" />
+          />
+          <Field
+            component={MarkdownField}
+            name="description"
+            label={t('Description')}
+            fullWidth={true}
+            multiline={true}
+            rows="4"
+            style={{ marginTop: 20 }}
+            onFocus={handleChangeFocus}
+            onSubmit={handleSubmitField}
+            helperText={
+              <SubscriptionFocus context={context} fieldName="description" />
               }
+          />
+          <Stack flexDirection="row" justifyContent="flex-end" mt={2}>
+            <Button
+              color="error"
+              variant="contained"
+              onClick={handleOpenDelete}
+              disabled={deleting}
+            >
+              {t('Delete')}
+            </Button>
+            <DeleteDialog
+              deletion={deletion}
+              submitDelete={submitDelete}
+              message={workspace.type === 'investigation'
+                ? t('Do you want to delete this investigation?')
+                : t('Do you want to delete this dashboard?')}
             />
-          </Form>
-        )}
-      </Formik>
-    );
-  }
-}
+          </Stack>
+        </Form>
+      )}
+    </Formik>
+  );
+};
 
 WorkspaceEditionOverviewComponent.propTypes = {
   classes: PropTypes.object,
